pyproject.toml
[tool.poetry]
name = "crm"
version = "0.1.0"
description = "Customer Relationship Management - sistema para controle de negociações em andamento"
authors = ["igor <igorjosafa16@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.2"
pytelegrambotapi = "^4.16.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
